crypto-encryption-app/requirements.txt
cryptography
pycryptodome
qiskit
pytest